+++ src/components/Login/LoginForm.js
@@ -9,31 +9,31 @@ import { I18N_KEYS } from '../../utils/constants';
 import style from '../../utils/FormControl/formsControls.module.css';
 
 const LoginForm = (props) => {
-    const { t } = useTranslation();
-    const { handleSubmit, error } = props;
+  const { t } = useTranslation();
+  const { handleSubmit, error } = props;
 
-    return (
-        <form onSubmit={handleSubmit}>
-            <div>
-                <Field placeholder={t(I18N_KEYS.EMAIL)} name="email" component={Input} validate={[required]} />
-            </div>
-            <div>
-                <Field placeholder={t(I18N_KEYS.PASSWORD)} name="password" component={Input} validate={[required]} />
-            </div>
-            <div>
-                <Field name="rememberMe" type="checkbox" component="input" />
-                <span>{t(I18N_KEYS.REMEMBER_ME)}</span>
-            </div>
-            {error && (
-                <div className={style.formSummaryError}>
-                    {error}
-                </div>
-            )}
-            <div>
-                <button>{t(I18N_KEYS.LOGIN)}</button>
-            </div>
-        </form>
-    );
+  return (
+    <form onSubmit={handleSubmit}>
+      <div>
+        <Field placeholder={t(I18N_KEYS.EMAIL)} name="email" component={Input} validate={[required]} />
+      </div>
+      <div>
+        <Field placeholder={t(I18N_KEYS.PASSWORD)} name="password" component={Input} validate={[required]} />
+      </div>
+      <div>
+        <Field name="rememberMe" type="checkbox" component="input" />
+        <span>{t(I18N_KEYS.REMEMBER_ME)}</span>
+      </div>
+      {error && (
+        <div className={style.formSummaryError}>
+          {error}
+        </div>
+      )}
+      <div>
+        <button>{t(I18N_KEYS.LOGIN)}</button>
+      </div>
+    </form>
+  );
 };
 
 const LoginReduxForm = reduxForm({ form: 'login' })(LoginForm);
