+++ src/components/Users/Users.module.css
@@ -2,7 +2,3 @@
     width: 100px;
     height: 100px;
 }
-
-.selectedPage {
-    font-weight: bold;
-}
