+++ src/utils/Paginator/paginator.module.css
@@ -1,0 +1,3 @@
+.selectedPage {
+    font-weight: bold;
+}
